+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added @phpstan-require-implements constraints
+
 ## v0.9.1 (2024-08-21)
 * Updated Lucid dependency